prefix not-for-export variables with underscore to help tab completion
this means that after import team_7564616d_interface, then using
team_7564616d_interface. <tab> will autocomplete to the function we want

this isn't best practice; ideally we'd want to use an __init__.py that
only exports the function we want, but that requires making more infrastructure
than a single-file wrapper.